requirements.txt
Django==1.9.6
django-cors-headers==2.4.1
django-rest-multiple-models==2.1.3
django-rest-swagger==2.1.0
django-wysiwyg-redactor==0.4.9
djangorestframework==3.5.3
gunicorn==19.6.0
Pillow
python-dateutil==2.6.0
pytz==2019.1
PyYAML==3.12
requests==2.12.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.1
watchdog==0.8.3
XlsxWriter==1.1.5

